Create read.me
This is a visual comparison of India w.r.t primarily China. 

I thought of creating an animated visualization inspired from famous TED talk of "Hans Rosling". This is visualization is in reference to the geo-political situation prevalent in India nowadays. Topic to ponder visually is – how a conflict impacts a nation’s growth? Secondly how each country grew in the last 70 years w.r.t to India? Can we see this in an animated manner?

I plotted based on publicly available data of the entire world. Keeping the rest countries aside, I concentrated on the growth of India vs China post 1940’s where in a new world order started to take shape. I have also included Australia, Japan, Pakistan, Israel and USA as these are also part of current discourse. Rest of the historical data is excluded.

Each bubble here is size of a nation based on its population for that particular year.